Adding Scraper base class to hold common logic between all scrapers
