+++ lib/vbulletin_scraper/V4/post_scraper.rb
@@ -1,10 +1,8 @@
-require 'nokogiri'
-require 'open-uri'
-require 'open_uri_redirections'
+require_relative 'scraper'
 
 module VbulletinScraper
     module V4
-        class PostScraper
+        class PostScraper < Scraper
             def initialize(input)
                 @data = nil
                 if input.start_with? "http" || "www"
@@ -89,51 +87,6 @@ module VbulletinScraper
                     return get_raw_text(permalink)
                 end
                 return ''
-            end
-                
-            def get_item_by_selector(selector)
-                if @data != nil
-                    if @data.at_css(selector)
-                        return @data.at_css(selector)
-                    end
-                end
-                return nil
-            end
-
-            def get_items_by_selector(selector)
-                if @data != nil
-                    if @data.css(selector)
-                        return @data.css(selector)
-                    end
-                end
-            end
-                
-            def get_item_by_selector_with_attribute(selector, attribute)
-                if @data != nil
-                    if @data.at_css(selector)
-                        return @data.at_css(selector)[attribute]
-                    end
-                end
-                return nil
-            end
-                
-            def get_raw_text(input)
-                if input != nil
-                    return input.strip.gsub(/\u00a0/, ' ').gsub('\t', '')
-                else
-                    return nil
-                end
-            end
-
-            def get_int(input)
-                if input != nil
-                    if input != ''
-                        begin
-                            return input.to_i
-                        end
-                    end
-                end        
-                return 0
             end
         end
     end
