+++ lib/vbulletin_scraper/V4/topic_scraper.rb
@@ -1,21 +1,8 @@
-require 'nokogiri'
-require 'open-uri'
-require 'open_uri_redirections'
+require_relative 'scraper'
 
 module VbulletinScraper
     module V4
-        class TopicScraper
-            def initialize(input)
-                @data = nil
-                if input.start_with? "http" || "www"
-                    @data = Nokogiri::HTML(open(input, :allow_redirections => :all))
-                    @data.encoding = "UTF-8"
-                else
-                    @data = Nokogiri::HTML(input)
-                    @data.encoding = "UTF-8"
-                end
-            end
-
+        class TopicScraper < Scraper
             def is_valid_vbulletin
                 if get_vbulletin_version != ''
                     return true
@@ -34,7 +21,7 @@ module VbulletinScraper
 
             def get_current_page_number
                 if is_valid_vbulletin
-                    pageNumber = get_item_by_selector('#pagination_top a.popupctrl')				
+                    pageNumber = get_item_by_selector('#pagination_top a.popupctrl')        
                     if pageNumber != nil
                         pageNumber = pageNumber.text.gsub('Page', '').gsub(' ', '').split('of').first
                         return get_int(get_raw_text(pageNumber))
@@ -94,51 +81,6 @@ module VbulletinScraper
                 else
                     return []
                 end
-            end
-
-            def get_item_by_selector(selector)
-                if @data != nil
-                    if @data.at_css(selector)
-                        return @data.at_css(selector)
-                    end
-                end
-                return nil
-            end
-
-            def get_items_by_selector(selector)
-                if @data != nil
-                    if @data.css(selector)
-                        return @data.css(selector)
-                    end
-                end
-            end
-                
-            def get_item_by_selector_with_attribute(selector, attribute)
-                if @data != nil
-                    if @data.at_css(selector)
-                        return @data.at_css(selector)[attribute]
-                    end
-                end
-                return nil
-            end
-                
-            def get_raw_text(input)
-                if input != nil
-                    return input.strip.gsub(/\u00a0/, ' ')
-                else
-                    return nil
-                end
-            end
-                
-            def get_int(input)
-                if input != nil
-                    if input != ''
-                        begin
-                            return input.to_i
-                        end
-                    end
-                end        
-                return 0
             end
         end
     end
